requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1600297518631/work
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.17.0
albucore==0.0.17
albumentations==0.3.2
annotated-types==0.7.0
anyio==3.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
ccimport==0.4.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clarabel==0.9.0
comm==0.2.2
coverage==7.6.1
cumm==0.4.11
cumm-cu113==0.4.11
cvxpy==1.5.2
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
docutils==0.21.2
ecos==2.0.14
entrypoints==0.4
erdos==0.3.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
filterpy==1.4.1
fire==0.7.0
flatbuffers==1.12
fqdn==1.5.1
future==1.0.0
gast==0.4.0
gdown==5.2.0
geohash2==1.1
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.10
imageio==2.35.1
imgaug==0.2.6
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_packaging==0.12.3
jupyter_server==2.14.2
jupyter_server_fileid==0.9.3
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lapsolver==1.1.0
lark==1.2.2
lazy_loader==0.4
llvmlite @ file:///croot/llvmlite_1697031035361/work
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.4.3
matplotlib-inline==0.1.7
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv==1.3.10
mmdet==2.14.0
motmetrics==1.1.3
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.2
ninja==1.11.1.3
nonechucks==0.3.1
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.19.2
nuscenes-devkit==1.1.1
oauthlib==3.2.2
open3d==0.13.0
opencv-contrib-python==4.4.0.42
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
osqp==0.6.7.post3
overrides==7.7.0
packaging==24.1
pandas==1.2.2
pandocfilters==1.5.1
parso==0.8.4
pccm==0.4.16
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.0.0
prettytable==3.11.0
progress==1.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycocotools==2.0.7
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==1.9.6
Pygments==2.18.0
pyparsing==3.1.4
pyquaternion==0.9.9
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post4
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.9
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
scs==3.2.7
Send2Trash==1.8.3
Shapely==1.6.4
six==1.15.0
sniffio==1.3.1
soupsieve==2.6
spconv==2.3.6
spconv-cu113==2.3.6
stack-data==0.6.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.1
termcolor==1.1.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.4.0
tomli==2.0.2
tomlkit==0.13.2
torch==1.4.0
torchvision==0.5.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
traj_dist==1.1
types-python-dateutil==2.9.0.20241003
typing-extensions==3.7.4
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
Wand==0.6.13
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.12.1
xmltodict==0.14.2
y-py==0.6.2
yapf==0.40.2
ypy-websocket==0.8.4
zipp @ file:///croot/zipp_1729012354496/work
